Split reqh.go into multiple files
The files are now sorted by specific tasks and thus are more organised.
